backend/pyproject.toml
[tool.poetry]
name = "GI-ChAn"
version = "0.2.1"
description = "Web-application for Genshin Impact characters' builds analyze."
authors = ["alex6712 <vanyukov.alex@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
fastapi = "^0.103.0"
python-multipart = "0.0.6"
email-validator = "^2.0.0.post2"
pydantic-settings = "^2.0.2"
pydantic-extra-types = "^2.0.0"
SQLAlchemy = "^2.0.19"
asyncpg = "^0.28.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.1.0" }
phonenumbers = "^8.13.18"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
pytest = "^7.4.0"
httpx = "^0.24.1"
trio = "^0.22.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
